+++ src/Components/ClickOutside/ClickOutside.stories.tsx
@@ -1,5 +1,6 @@
 // Libraries
 import * as React from 'react'
+import marked from 'marked'
 
 // Storybook
 import {storiesOf} from '@storybook/react'
@@ -8,15 +9,26 @@ import {jsxDecorator} from 'storybook-addon-jsx'
 // Components
 import {ClickOutside} from './ClickOutside'
 
+// Notes
+const ClickOutsideReadme = marked(require('./ClickOutside.md'))
+
 const clickOutsideStories = storiesOf(
   'Utilities|ClickOutside',
   module
 ).addDecorator(jsxDecorator)
 
-clickOutsideStories.add('Example', () => {
-  return (
-    <ClickOutside onClickOutside={() => alert('Clicked outside')}>
-      <div className="mockComponent box" />
-    </ClickOutside>
-  )
-})
+clickOutsideStories.add(
+  'Example',
+  () => (
+    <div className="story--example">
+      <ClickOutside onClickOutside={() => alert('Clicked outside')}>
+        <div className="mockComponent box">Click outside this element</div>
+      </ClickOutside>
+    </div>
+  ),
+  {
+    readme: {
+      content: ClickOutsideReadme,
+    },
+  }
+)